Refactor The Home Ui Widgets
