+++ lib/features/home/ui/widgets/doctors_list/doctor_list_view_card.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import '../../../../core/helpers/assets.dart';
-import '../../../../core/helpers/spacing.dart';
-import '../../../../core/themes/colors.dart';
-import '../../../../core/themes/text_styles.dart';
-import '../../data/models/specializations_response_model.dart';
+import '../../../../../core/helpers/assets.dart';
+import '../../../../../core/helpers/spacing.dart';
+import '../../../../../core/themes/colors.dart';
+import '../../../../../core/themes/text_styles.dart';
+import '../../../data/models/specializations_response_model.dart';
 
 class DoctorListViewCard extends StatelessWidget {
   const DoctorListViewCard({super.key, required this.doctorModel});
